Refactor daemon implementation into modular sections
