+++ crates/daemon/src/daemon/sections/config_parsing.rs
@@ -1,0 +1,582 @@
+#[derive(Clone, Debug, Eq, PartialEq)]
+struct ConfigDirectiveOrigin {
+    path: PathBuf,
+    line: usize,
+}
+
+#[derive(Debug)]
+struct ParsedConfigModules {
+    modules: Vec<ModuleDefinition>,
+    global_refuse_options: Vec<(Vec<String>, ConfigDirectiveOrigin)>,
+    motd_lines: Vec<String>,
+    pid_file: Option<(PathBuf, ConfigDirectiveOrigin)>,
+    reverse_lookup: Option<(bool, ConfigDirectiveOrigin)>,
+    lock_file: Option<(PathBuf, ConfigDirectiveOrigin)>,
+    global_bandwidth_limit: Option<(BandwidthLimitComponents, ConfigDirectiveOrigin)>,
+    global_secrets_file: Option<(PathBuf, ConfigDirectiveOrigin)>,
+}
+
+fn parse_config_modules(path: &Path) -> Result<ParsedConfigModules, DaemonError> {
+    let mut stack = Vec::new();
+    parse_config_modules_inner(path, &mut stack)
+}
+
+fn parse_config_modules_inner(
+    path: &Path,
+    stack: &mut Vec<PathBuf>,
+) -> Result<ParsedConfigModules, DaemonError> {
+    let canonical = path
+        .canonicalize()
+        .map_err(|error| config_io_error("read", path, error))?;
+
+    if stack.iter().any(|seen| seen == &canonical) {
+        return Err(config_parse_error(
+            path,
+            0,
+            format!("recursive include detected for '{}'", canonical.display()),
+        ));
+    }
+
+    let contents = fs::read_to_string(&canonical)
+        .map_err(|error| config_io_error("read", &canonical, error))?;
+    stack.push(canonical.clone());
+
+    let mut modules = Vec::new();
+    let mut current: Option<ModuleDefinitionBuilder> = None;
+    let mut global_refuse_directives = Vec::new();
+    let mut global_refuse_line: Option<usize> = None;
+    let mut motd_lines = Vec::new();
+    let mut pid_file: Option<(PathBuf, ConfigDirectiveOrigin)> = None;
+    let mut reverse_lookup: Option<(bool, ConfigDirectiveOrigin)> = None;
+    let mut lock_file: Option<(PathBuf, ConfigDirectiveOrigin)> = None;
+    let mut global_bwlimit: Option<(BandwidthLimitComponents, ConfigDirectiveOrigin)> = None;
+    let mut global_secrets_file: Option<(PathBuf, ConfigDirectiveOrigin)> = None;
+
+    let result = (|| -> Result<ParsedConfigModules, DaemonError> {
+        for (index, raw_line) in contents.lines().enumerate() {
+            let line_number = index + 1;
+            let line = raw_line.trim();
+
+            if line.is_empty() || line.starts_with('#') || line.starts_with(';') {
+                continue;
+            }
+
+            if line.starts_with('[') {
+                let end = line.find(']').ok_or_else(|| {
+                    config_parse_error(path, line_number, "unterminated module header")
+                })?;
+                let name = line[1..end].trim();
+
+                if name.is_empty() {
+                    return Err(config_parse_error(
+                        path,
+                        line_number,
+                        "module name must be non-empty",
+                    ));
+                }
+
+                ensure_valid_module_name(name)
+                    .map_err(|msg| config_parse_error(path, line_number, msg))?;
+
+                let trailing = line[end + 1..].trim();
+                if !trailing.is_empty() && !trailing.starts_with('#') && !trailing.starts_with(';')
+                {
+                    return Err(config_parse_error(
+                        path,
+                        line_number,
+                        "unexpected characters after module header",
+                    ));
+                }
+
+                if let Some(builder) = current.take() {
+                    let default_secrets = global_secrets_file.as_ref().map(|(p, _)| p.as_path());
+                    modules.push(builder.finish(path, default_secrets)?);
+                }
+
+                current = Some(ModuleDefinitionBuilder::new(name.to_string(), line_number));
+                continue;
+            }
+
+            let (key, value) = line.split_once('=').ok_or_else(|| {
+                config_parse_error(path, line_number, "expected 'key = value' directive")
+            })?;
+            let key = key.trim().to_ascii_lowercase();
+            let value = value.trim();
+
+            if let Some(builder) = current.as_mut() {
+                match key.as_str() {
+                    "path" => {
+                        if value.is_empty() {
+                            return Err(config_parse_error(
+                                path,
+                                line_number,
+                                "module path directive must not be empty",
+                            ));
+                        }
+                        builder.set_path(PathBuf::from(value), path, line_number)?;
+                    }
+                    "comment" => {
+                        let comment = if value.is_empty() {
+                            None
+                        } else {
+                            Some(value.to_string())
+                        };
+                        builder.set_comment(comment, path, line_number)?;
+                    }
+                    "hosts allow" => {
+                        let patterns = parse_host_list(value, path, line_number, "hosts allow")?;
+                        builder.set_hosts_allow(patterns, path, line_number)?;
+                    }
+                    "hosts deny" => {
+                        let patterns = parse_host_list(value, path, line_number, "hosts deny")?;
+                        builder.set_hosts_deny(patterns, path, line_number)?;
+                    }
+                    "auth users" => {
+                        let users = parse_auth_user_list(value).map_err(|error| {
+                            config_parse_error(
+                                path,
+                                line_number,
+                                format!("invalid 'auth users' directive: {error}"),
+                            )
+                        })?;
+                        builder.set_auth_users(users, path, line_number)?;
+                    }
+                    "secrets file" => {
+                        if value.is_empty() {
+                            return Err(config_parse_error(
+                                path,
+                                line_number,
+                                "'secrets file' directive must not be empty",
+                            ));
+                        }
+                        builder.set_secrets_file(PathBuf::from(value), path, line_number)?;
+                    }
+                    "bwlimit" => {
+                        if value.is_empty() {
+                            return Err(config_parse_error(
+                                path,
+                                line_number,
+                                "'bwlimit' directive must not be empty",
+                            ));
+                        }
+                        let components = parse_config_bwlimit(value, path, line_number)?;
+                        builder.set_bandwidth_limit(
+                            components.rate(),
+                            components.burst(),
+                            components.burst_specified(),
+                            path,
+                            line_number,
+                        )?;
+                    }
+                    "refuse options" => {
+                        let options = parse_refuse_option_list(value).map_err(|error| {
+                            config_parse_error(
+                                path,
+                                line_number,
+                                format!("invalid 'refuse options' directive: {error}"),
+                            )
+                        })?;
+                        builder.set_refuse_options(options, path, line_number)?;
+                    }
+                    "read only" => {
+                        let parsed = parse_boolean_directive(value).ok_or_else(|| {
+                            config_parse_error(
+                                path,
+                                line_number,
+                                format!("invalid boolean value '{value}' for 'read only'"),
+                            )
+                        })?;
+                        builder.set_read_only(parsed, path, line_number)?;
+                    }
+                    "use chroot" => {
+                        let parsed = parse_boolean_directive(value).ok_or_else(|| {
+                            config_parse_error(
+                                path,
+                                line_number,
+                                format!("invalid boolean value '{value}' for 'use chroot'"),
+                            )
+                        })?;
+                        builder.set_use_chroot(parsed, path, line_number)?;
+                    }
+                    "numeric ids" => {
+                        let parsed = parse_boolean_directive(value).ok_or_else(|| {
+                            config_parse_error(
+                                path,
+                                line_number,
+                                format!("invalid boolean value '{value}' for 'numeric ids'"),
+                            )
+                        })?;
+                        builder.set_numeric_ids(parsed, path, line_number)?;
+                    }
+                    "list" => {
+                        let parsed = parse_boolean_directive(value).ok_or_else(|| {
+                            config_parse_error(
+                                path,
+                                line_number,
+                                format!("invalid boolean value '{value}' for 'list'"),
+                            )
+                        })?;
+                        builder.set_listable(parsed, path, line_number)?;
+                    }
+                    "uid" => {
+                        let uid = parse_numeric_identifier(value).ok_or_else(|| {
+                            config_parse_error(path, line_number, format!("invalid uid '{value}'"))
+                        })?;
+                        builder.set_uid(uid, path, line_number)?;
+                    }
+                    "gid" => {
+                        let gid = parse_numeric_identifier(value).ok_or_else(|| {
+                            config_parse_error(path, line_number, format!("invalid gid '{value}'"))
+                        })?;
+                        builder.set_gid(gid, path, line_number)?;
+                    }
+                    "timeout" => {
+                        let timeout = parse_timeout_seconds(value).ok_or_else(|| {
+                            config_parse_error(
+                                path,
+                                line_number,
+                                format!("invalid timeout '{value}'"),
+                            )
+                        })?;
+                        builder.set_timeout(timeout, path, line_number)?;
+                    }
+                    "max connections" => {
+                        let max = parse_max_connections_directive(value).ok_or_else(|| {
+                            config_parse_error(
+                                path,
+                                line_number,
+                                format!("invalid max connections value '{value}'"),
+                            )
+                        })?;
+                        builder.set_max_connections(max, path, line_number)?;
+                    }
+                    _ => {
+                        // Unsupported directives are ignored for now.
+                    }
+                }
+                continue;
+            }
+
+            match key.as_str() {
+                "refuse options" => {
+                    if value.is_empty() {
+                        return Err(config_parse_error(
+                            path,
+                            line_number,
+                            "'refuse options' directive must not be empty",
+                        ));
+                    }
+                    let options = parse_refuse_option_list(value).map_err(|error| {
+                        config_parse_error(
+                            path,
+                            line_number,
+                            format!("invalid 'refuse options' directive: {error}"),
+                        )
+                    })?;
+
+                    if let Some(existing_line) = global_refuse_line {
+                        return Err(config_parse_error(
+                            path,
+                            line_number,
+                            format!(
+                                "duplicate 'refuse options' directive in global section (previously defined on line {})",
+                                existing_line
+                            ),
+                        ));
+                    }
+
+                    global_refuse_line = Some(line_number);
+                    global_refuse_directives.push((
+                        options,
+                        ConfigDirectiveOrigin {
+                            path: canonical.clone(),
+                            line: line_number,
+                        },
+                    ));
+                }
+                "include" => {
+                    let trimmed = value.trim();
+                    if trimmed.is_empty() {
+                        return Err(config_parse_error(
+                            path,
+                            line_number,
+                            "'include' directive must not be empty",
+                        ));
+                    }
+
+                    let include_path = resolve_config_relative_path(&canonical, trimmed);
+                    let included = parse_config_modules_inner(&include_path, stack)?;
+
+                    if !included.modules.is_empty() {
+                        modules.extend(included.modules);
+                    }
+
+                    if !included.motd_lines.is_empty() {
+                        motd_lines.extend(included.motd_lines);
+                    }
+
+                    if !included.global_refuse_options.is_empty() {
+                        global_refuse_directives.extend(included.global_refuse_options);
+                    }
+
+                    if let Some((components, origin)) = included.global_bandwidth_limit {
+                        if let Some((existing, existing_origin)) = &global_bwlimit {
+                            if existing != &components {
+                                return Err(config_parse_error(
+                                    &origin.path,
+                                    origin.line,
+                                    format!(
+                                        "duplicate 'bwlimit' directive in global section (previously defined on line {})",
+                                        existing_origin.line
+                                    ),
+                                ));
+                            }
+                        } else {
+                            global_bwlimit = Some((components, origin));
+                        }
+                    }
+
+                    if let Some((secrets_path, origin)) = included.global_secrets_file {
+                        if let Some((existing, existing_origin)) = &global_secrets_file {
+                            if existing != &secrets_path {
+                                return Err(config_parse_error(
+                                    &origin.path,
+                                    origin.line,
+                                    format!(
+                                        "duplicate 'secrets file' directive in global section (previously defined on line {})",
+                                        existing_origin.line
+                                    ),
+                                ));
+                            }
+                        } else {
+                            global_secrets_file = Some((secrets_path, origin));
+                        }
+                    }
+                }
+                "motd file" => {
+                    let trimmed = value.trim();
+                    if trimmed.is_empty() {
+                        return Err(config_parse_error(
+                            path,
+                            line_number,
+                            "'motd file' directive must not be empty",
+                        ));
+                    }
+
+                    let motd_path = resolve_config_relative_path(path, trimmed);
+                    let contents = fs::read_to_string(&motd_path).map_err(|error| {
+                        config_parse_error(
+                            path,
+                            line_number,
+                            format!(
+                                "failed to read motd file '{}': {}",
+                                motd_path.display(),
+                                error
+                            ),
+                        )
+                    })?;
+
+                    for raw_line in contents.lines() {
+                        motd_lines.push(raw_line.trim_end_matches('\r').to_string());
+                    }
+                }
+                "motd" => {
+                    motd_lines.push(value.trim_end_matches(['\r', '\n']).to_string());
+                }
+                "pid file" => {
+                    let trimmed = value.trim();
+                    if trimmed.is_empty() {
+                        return Err(config_parse_error(
+                            path,
+                            line_number,
+                            "'pid file' directive must not be empty",
+                        ));
+                    }
+
+                    let resolved = resolve_config_relative_path(path, trimmed);
+                    if let Some((existing, origin)) = &pid_file {
+                        if existing != &resolved {
+                            return Err(config_parse_error(
+                                path,
+                                line_number,
+                                format!(
+                                    "duplicate 'pid file' directive in global section (previously defined on line {})",
+                                    origin.line
+                                ),
+                            ));
+                        }
+                    } else {
+                        pid_file = Some((
+                            resolved,
+                            ConfigDirectiveOrigin {
+                                path: canonical.clone(),
+                                line: line_number,
+                            },
+                        ));
+                    }
+                }
+                "reverse lookup" => {
+                    let parsed = parse_boolean_directive(value).ok_or_else(|| {
+                        config_parse_error(
+                            path,
+                            line_number,
+                            format!("invalid boolean value '{}' for 'reverse lookup'", value),
+                        )
+                    })?;
+
+                    let origin = ConfigDirectiveOrigin {
+                        path: canonical.clone(),
+                        line: line_number,
+                    };
+
+                    if let Some((existing, existing_origin)) = &reverse_lookup {
+                        if *existing != parsed {
+                            return Err(config_parse_error(
+                                path,
+                                line_number,
+                                format!(
+                                    "duplicate 'reverse lookup' directive in global section (previously defined on line {})",
+                                    existing_origin.line
+                                ),
+                            ));
+                        }
+                    } else {
+                        reverse_lookup = Some((parsed, origin));
+                    }
+                }
+                "bwlimit" => {
+                    if value.is_empty() {
+                        return Err(config_parse_error(
+                            path,
+                            line_number,
+                            "'bwlimit' directive must not be empty",
+                        ));
+                    }
+
+                    let components = parse_config_bwlimit(value, path, line_number)?;
+                    let origin = ConfigDirectiveOrigin {
+                        path: canonical.clone(),
+                        line: line_number,
+                    };
+
+                    if let Some((existing, existing_origin)) = &global_bwlimit {
+                        if existing != &components {
+                            return Err(config_parse_error(
+                                &origin.path,
+                                origin.line,
+                                format!(
+                                    "duplicate 'bwlimit' directive in global section (previously defined on line {})",
+                                    existing_origin.line
+                                ),
+                            ));
+                        }
+                    } else {
+                        global_bwlimit = Some((components, origin));
+                    }
+                }
+                "secrets file" => {
+                    let trimmed = value.trim();
+                    if trimmed.is_empty() {
+                        return Err(config_parse_error(
+                            path,
+                            line_number,
+                            "'secrets file' directive must not be empty",
+                        ));
+                    }
+
+                    let resolved = resolve_config_relative_path(path, trimmed);
+                    let validated = validate_secrets_file(&resolved, path, line_number)?;
+                    let origin = ConfigDirectiveOrigin {
+                        path: canonical.clone(),
+                        line: line_number,
+                    };
+
+                    if let Some((existing, existing_origin)) = &global_secrets_file {
+                        if existing != &validated {
+                            return Err(config_parse_error(
+                                path,
+                                line_number,
+                                format!(
+                                    "duplicate 'secrets file' directive in global section (previously defined on line {})",
+                                    existing_origin.line
+                                ),
+                            ));
+                        }
+                    } else {
+                        global_secrets_file = Some((validated, origin));
+                    }
+                }
+                "lock file" => {
+                    let trimmed = value.trim();
+                    if trimmed.is_empty() {
+                        return Err(config_parse_error(
+                            path,
+                            line_number,
+                            "'lock file' directive must not be empty",
+                        ));
+                    }
+
+                    let resolved = resolve_config_relative_path(path, trimmed);
+                    let origin = ConfigDirectiveOrigin {
+                        path: canonical.clone(),
+                        line: line_number,
+                    };
+
+                    if let Some((existing, existing_origin)) = &lock_file {
+                        if existing != &resolved {
+                            return Err(config_parse_error(
+                                path,
+                                line_number,
+                                format!(
+                                    "duplicate 'lock file' directive in global section (previously defined on line {})",
+                                    existing_origin.line
+                                ),
+                            ));
+                        }
+                    } else {
+                        lock_file = Some((resolved, origin));
+                    }
+                }
+                _ => {
+                    return Err(config_parse_error(
+                        path,
+                        line_number,
+                        "directive outside module section",
+                    ));
+                }
+            }
+        }
+
+        if let Some(builder) = current {
+            let default_secrets = global_secrets_file.as_ref().map(|(p, _)| p.as_path());
+            modules.push(builder.finish(path, default_secrets)?);
+        }
+
+        Ok(ParsedConfigModules {
+            modules,
+            global_refuse_options: global_refuse_directives,
+            motd_lines,
+            pid_file,
+            reverse_lookup,
+            lock_file,
+            global_bandwidth_limit: global_bwlimit,
+            global_secrets_file,
+        })
+    })();
+
+    stack.pop();
+    result
+}
+
+fn resolve_config_relative_path(config_path: &Path, value: &str) -> PathBuf {
+    let candidate = Path::new(value);
+    if candidate.is_absolute() {
+        return candidate.to_path_buf();
+    }
+
+    if let Some(parent) = config_path.parent() {
+        parent.join(candidate)
+    } else {
+        candidate.to_path_buf()
+    }
+}
