+++ crates/daemon/src/daemon/sections/module_access.rs
@@ -1,0 +1,524 @@
+fn respond_with_module_list(
+    stream: &mut TcpStream,
+    limiter: &mut Option<BandwidthLimiter>,
+    modules: &[ModuleRuntime],
+    motd_lines: &[String],
+    peer_ip: IpAddr,
+    reverse_lookup: bool,
+) -> io::Result<()> {
+    for line in motd_lines {
+        let payload = if line.is_empty() {
+            "MOTD".to_string()
+        } else {
+            format!("MOTD {line}")
+        };
+        let message = format_legacy_daemon_message(LegacyDaemonMessage::Other(&payload));
+        write_limited(stream, limiter, message.as_bytes())?;
+    }
+
+    let ok = format_legacy_daemon_message(LegacyDaemonMessage::Ok);
+    write_limited(stream, limiter, ok.as_bytes())?;
+
+    let mut hostname_cache: Option<Option<String>> = None;
+    for module in modules {
+        if !module.listable {
+            continue;
+        }
+
+        let peer_host = module_peer_hostname(module, &mut hostname_cache, peer_ip, reverse_lookup);
+        if !module.permits(peer_ip, peer_host) {
+            continue;
+        }
+
+        let mut line = module.name.clone();
+        if let Some(comment) = &module.comment {
+            if !comment.is_empty() {
+                line.push('\t');
+                line.push_str(comment);
+            }
+        }
+        line.push('\n');
+        write_limited(stream, limiter, line.as_bytes())?;
+    }
+
+    let exit = format_legacy_daemon_message(LegacyDaemonMessage::Exit);
+    write_limited(stream, limiter, exit.as_bytes())?;
+    stream.flush()
+}
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+enum AuthenticationStatus {
+    Granted,
+    Denied,
+}
+
+fn perform_module_authentication(
+    reader: &mut BufReader<TcpStream>,
+    limiter: &mut Option<BandwidthLimiter>,
+    module: &ModuleDefinition,
+    peer_ip: IpAddr,
+) -> io::Result<AuthenticationStatus> {
+    let challenge = generate_auth_challenge(peer_ip);
+    {
+        let stream = reader.get_mut();
+        let message = format_legacy_daemon_message(LegacyDaemonMessage::AuthRequired {
+            module: Some(&challenge),
+        });
+        write_limited(stream, limiter, message.as_bytes())?;
+        stream.flush()?;
+    }
+
+    let response = match read_trimmed_line(reader)? {
+        Some(line) => line,
+        None => {
+            deny_module(reader.get_mut(), module, peer_ip, limiter)?;
+            return Ok(AuthenticationStatus::Denied);
+        }
+    };
+
+    let mut segments = response.splitn(2, |ch: char| ch.is_ascii_whitespace());
+    let username = segments.next().unwrap_or_default();
+    let digest = segments
+        .next()
+        .map(|segment| segment.trim_start_matches(|ch: char| ch.is_ascii_whitespace()))
+        .unwrap_or("");
+
+    if username.is_empty() || digest.is_empty() {
+        deny_module(reader.get_mut(), module, peer_ip, limiter)?;
+        return Ok(AuthenticationStatus::Denied);
+    }
+
+    if !module.auth_users.iter().any(|user| user == username) {
+        deny_module(reader.get_mut(), module, peer_ip, limiter)?;
+        return Ok(AuthenticationStatus::Denied);
+    }
+
+    if !verify_secret_response(module, username, &challenge, digest)? {
+        deny_module(reader.get_mut(), module, peer_ip, limiter)?;
+        return Ok(AuthenticationStatus::Denied);
+    }
+
+    Ok(AuthenticationStatus::Granted)
+}
+
+fn generate_auth_challenge(peer_ip: IpAddr) -> String {
+    let mut input = [0u8; 32];
+    let address_text = peer_ip.to_string();
+    let address_bytes = address_text.as_bytes();
+    let copy_len = address_bytes.len().min(16);
+    input[..copy_len].copy_from_slice(&address_bytes[..copy_len]);
+
+    let timestamp = SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .unwrap_or_default();
+    let seconds = (timestamp.as_secs() & u64::from(u32::MAX)) as u32;
+    let micros = timestamp.subsec_micros();
+    let pid = std::process::id();
+
+    input[16..20].copy_from_slice(&seconds.to_le_bytes());
+    input[20..24].copy_from_slice(&micros.to_le_bytes());
+    input[24..28].copy_from_slice(&pid.to_le_bytes());
+
+    let mut hasher = Md5::new();
+    hasher.update(&input);
+    let digest = hasher.finalize();
+    STANDARD_NO_PAD.encode(digest)
+}
+
+fn verify_secret_response(
+    module: &ModuleDefinition,
+    username: &str,
+    challenge: &str,
+    response: &str,
+) -> io::Result<bool> {
+    let secrets_path = match &module.secrets_file {
+        Some(path) => path,
+        None => return Ok(false),
+    };
+
+    let contents = fs::read_to_string(secrets_path)?;
+
+    for raw_line in contents.lines() {
+        let line = raw_line.trim_end_matches('\r');
+        if line.is_empty() || line.starts_with('#') {
+            continue;
+        }
+
+        if let Some((user, secret)) = line.split_once(':') {
+            if user == username {
+                let expected = compute_auth_response(secret, challenge);
+                return Ok(expected == response);
+            }
+        }
+    }
+
+    Ok(false)
+}
+
+fn compute_auth_response(secret: &str, challenge: &str) -> String {
+    let mut hasher = Md5::new();
+    hasher.update(secret.as_bytes());
+    hasher.update(challenge.as_bytes());
+    let digest = hasher.finalize();
+    STANDARD_NO_PAD.encode(digest)
+}
+
+fn deny_module(
+    stream: &mut TcpStream,
+    module: &ModuleDefinition,
+    peer_ip: IpAddr,
+    limiter: &mut Option<BandwidthLimiter>,
+) -> io::Result<()> {
+    let module_display = sanitize_module_identifier(&module.name);
+    let payload = ACCESS_DENIED_PAYLOAD
+        .replace("{module}", module_display.as_ref())
+        .replace("{addr}", &peer_ip.to_string());
+    write_limited(stream, limiter, payload.as_bytes())?;
+    write_limited(stream, limiter, b"\n")?;
+    let exit = format_legacy_daemon_message(LegacyDaemonMessage::Exit);
+    write_limited(stream, limiter, exit.as_bytes())?;
+    stream.flush()
+}
+
+fn send_daemon_ok(
+    stream: &mut TcpStream,
+    limiter: &mut Option<BandwidthLimiter>,
+) -> io::Result<()> {
+    let ok = format_legacy_daemon_message(LegacyDaemonMessage::Ok);
+    write_limited(stream, limiter, ok.as_bytes())?;
+    stream.flush()
+}
+
+/// Applies the module-specific bandwidth directives to the active limiter.
+///
+/// The helper mirrors upstream rsync's precedence rules: a module `bwlimit`
+/// directive overrides the daemon-wide limit with the strictest rate while
+/// honouring explicitly configured bursts. When a module omits the directive
+/// the limiter remains in the state established by the daemon scope, ensuring
+/// clients observe inherited throttling exactly as the C implementation does.
+/// The function returns the [`LimiterChange`] reported by
+/// [`apply_effective_limit`], allowing callers and tests to verify whether the
+/// limiter configuration changed as a result of the module overrides.
+fn apply_module_bandwidth_limit(
+    limiter: &mut Option<BandwidthLimiter>,
+    module_limit: Option<NonZeroU64>,
+    module_limit_specified: bool,
+    module_limit_configured: bool,
+    module_burst: Option<NonZeroU64>,
+    module_burst_specified: bool,
+) -> LimiterChange {
+    if module_limit_configured && module_limit.is_none() {
+        let burst_only_override =
+            module_burst_specified && module_burst.is_some() && limiter.is_some();
+        if !burst_only_override {
+            return if limiter.take().is_some() {
+                LimiterChange::Disabled
+            } else {
+                LimiterChange::Unchanged
+            };
+        }
+    }
+
+    let limit_specified =
+        module_limit_specified || (module_limit_configured && module_limit.is_some());
+    let burst_specified =
+        module_burst_specified && (module_limit_configured || module_limit_specified);
+
+    BandwidthLimitComponents::new_with_flags(
+        module_limit,
+        module_burst,
+        limit_specified,
+        burst_specified,
+    )
+    .apply_to_limiter(limiter)
+}
+
+#[allow(clippy::too_many_arguments)]
+fn respond_with_module_request(
+    reader: &mut BufReader<TcpStream>,
+    limiter: &mut Option<BandwidthLimiter>,
+    modules: &[ModuleRuntime],
+    request: &str,
+    peer_ip: IpAddr,
+    session_peer_host: Option<&str>,
+    options: &[String],
+    log_sink: Option<&SharedLogSink>,
+    reverse_lookup: bool,
+) -> io::Result<()> {
+    if let Some(module) = modules.iter().find(|module| module.name == request) {
+        let change = apply_module_bandwidth_limit(
+            limiter,
+            module.bandwidth_limit(),
+            module.bandwidth_limit_specified(),
+            module.bandwidth_limit_configured(),
+            module.bandwidth_burst(),
+            module.bandwidth_burst_specified(),
+        );
+
+        let mut hostname_cache: Option<Option<String>> = None;
+        let module_peer_host =
+            module_peer_hostname(module, &mut hostname_cache, peer_ip, reverse_lookup);
+
+        if change != LimiterChange::Unchanged {
+            if let Some(log) = log_sink {
+                log_module_bandwidth_change(
+                    log,
+                    module_peer_host.or(session_peer_host),
+                    peer_ip,
+                    request,
+                    limiter.as_ref(),
+                    change,
+                );
+            }
+        }
+        if module.permits(peer_ip, module_peer_host) {
+            let _connection_guard = match module.try_acquire_connection() {
+                Ok(guard) => guard,
+                Err(ModuleConnectionError::Limit(limit)) => {
+                    let payload =
+                        MODULE_MAX_CONNECTIONS_PAYLOAD.replace("{limit}", &limit.get().to_string());
+                    let stream = reader.get_mut();
+                    write_limited(stream, limiter, payload.as_bytes())?;
+                    write_limited(stream, limiter, b"\n")?;
+                    let exit = format_legacy_daemon_message(LegacyDaemonMessage::Exit);
+                    write_limited(stream, limiter, exit.as_bytes())?;
+                    stream.flush()?;
+                    if let Some(log) = log_sink {
+                        log_module_limit(
+                            log,
+                            module_peer_host.or(session_peer_host),
+                            peer_ip,
+                            request,
+                            limit,
+                        );
+                    }
+                    return Ok(());
+                }
+                Err(ModuleConnectionError::Io(error)) => {
+                    let stream = reader.get_mut();
+                    write_limited(stream, limiter, MODULE_LOCK_ERROR_PAYLOAD.as_bytes())?;
+                    write_limited(stream, limiter, b"\n")?;
+                    let exit = format_legacy_daemon_message(LegacyDaemonMessage::Exit);
+                    write_limited(stream, limiter, exit.as_bytes())?;
+                    stream.flush()?;
+                    if let Some(log) = log_sink {
+                        log_module_lock_error(
+                            log,
+                            module_peer_host.or(session_peer_host),
+                            peer_ip,
+                            request,
+                            &error,
+                        );
+                    }
+                    return Ok(());
+                }
+            };
+
+            if let Some(log) = log_sink {
+                log_module_request(
+                    log,
+                    module_peer_host.or(session_peer_host),
+                    peer_ip,
+                    request,
+                );
+            }
+
+            if let Some(refused) = refused_option(module, options) {
+                let payload = format!("@ERROR: The server is configured to refuse {}", refused);
+                let stream = reader.get_mut();
+                write_limited(stream, limiter, payload.as_bytes())?;
+                write_limited(stream, limiter, b"\n")?;
+                let exit = format_legacy_daemon_message(LegacyDaemonMessage::Exit);
+                write_limited(stream, limiter, exit.as_bytes())?;
+                stream.flush()?;
+                if let Some(log) = log_sink {
+                    log_module_refused_option(
+                        log,
+                        module_peer_host.or(session_peer_host),
+                        peer_ip,
+                        request,
+                        refused,
+                    );
+                }
+                return Ok(());
+            }
+
+            apply_module_timeout(reader.get_mut(), module)?;
+            let mut acknowledged = false;
+            if module.requires_authentication() {
+                match perform_module_authentication(reader, limiter, module, peer_ip)? {
+                    AuthenticationStatus::Denied => {
+                        if let Some(log) = log_sink {
+                            log_module_auth_failure(
+                                log,
+                                module_peer_host.or(session_peer_host),
+                                peer_ip,
+                                request,
+                            );
+                        }
+                        return Ok(());
+                    }
+                    AuthenticationStatus::Granted => {
+                        if let Some(log) = log_sink {
+                            log_module_auth_success(
+                                log,
+                                module_peer_host.or(session_peer_host),
+                                peer_ip,
+                                request,
+                            );
+                        }
+                        send_daemon_ok(reader.get_mut(), limiter)?;
+                        acknowledged = true;
+                    }
+                }
+            }
+
+            if !acknowledged {
+                send_daemon_ok(reader.get_mut(), limiter)?;
+            }
+
+            let module_display = sanitize_module_identifier(request);
+            let payload = MODULE_UNAVAILABLE_PAYLOAD.replace("{module}", module_display.as_ref());
+            let stream = reader.get_mut();
+            write_limited(stream, limiter, payload.as_bytes())?;
+            write_limited(stream, limiter, b"\n")?;
+            if let Some(log) = log_sink {
+                log_module_unavailable(
+                    log,
+                    module_peer_host.or(session_peer_host),
+                    peer_ip,
+                    request,
+                );
+            }
+        } else {
+            if let Some(log) = log_sink {
+                log_module_denied(
+                    log,
+                    module_peer_host.or(session_peer_host),
+                    peer_ip,
+                    request,
+                );
+            }
+            deny_module(reader.get_mut(), module, peer_ip, limiter)?;
+            return Ok(());
+        }
+    } else {
+        let module_display = sanitize_module_identifier(request);
+        let payload = UNKNOWN_MODULE_PAYLOAD.replace("{module}", module_display.as_ref());
+        let stream = reader.get_mut();
+        write_limited(stream, limiter, payload.as_bytes())?;
+        write_limited(stream, limiter, b"\n")?;
+        if let Some(log) = log_sink {
+            log_unknown_module(log, session_peer_host, peer_ip, request);
+        }
+    }
+
+    let exit = format_legacy_daemon_message(LegacyDaemonMessage::Exit);
+    let stream = reader.get_mut();
+    write_limited(stream, limiter, exit.as_bytes())?;
+    stream.flush()
+}
+
+fn open_log_sink(path: &Path) -> Result<SharedLogSink, DaemonError> {
+    let file = OpenOptions::new()
+        .create(true)
+        .append(true)
+        .open(path)
+        .map_err(|error| log_file_error(path, error))?;
+    Ok(Arc::new(Mutex::new(MessageSink::new(file))))
+}
+
+fn log_file_error(path: &Path, error: io::Error) -> DaemonError {
+    DaemonError::new(
+        FEATURE_UNAVAILABLE_EXIT_CODE,
+        rsync_error!(
+            FEATURE_UNAVAILABLE_EXIT_CODE,
+            format!("failed to open log file '{}': {}", path.display(), error)
+        )
+        .with_role(Role::Daemon),
+    )
+}
+
+fn pid_file_error(path: &Path, error: io::Error) -> DaemonError {
+    DaemonError::new(
+        FEATURE_UNAVAILABLE_EXIT_CODE,
+        rsync_error!(
+            FEATURE_UNAVAILABLE_EXIT_CODE,
+            format!("failed to write pid file '{}': {}", path.display(), error)
+        )
+        .with_role(Role::Daemon),
+    )
+}
+
+fn lock_file_error(path: &Path, error: io::Error) -> DaemonError {
+    DaemonError::new(
+        FEATURE_UNAVAILABLE_EXIT_CODE,
+        rsync_error!(
+            FEATURE_UNAVAILABLE_EXIT_CODE,
+            format!("failed to open lock file '{}': {}", path.display(), error)
+        )
+        .with_role(Role::Daemon),
+    )
+}
+
+fn log_message(log: &SharedLogSink, message: &Message) {
+    if let Ok(mut sink) = log.lock() {
+        if sink.write(message).is_ok() {
+            let _ = sink.flush();
+        }
+    }
+}
+
+fn format_host(host: Option<&str>, fallback: IpAddr) -> String {
+    host.map(str::to_string)
+        .unwrap_or_else(|| fallback.to_string())
+}
+
+/// Returns a sanitised view of a module identifier suitable for diagnostics.
+///
+/// Module names originate from user input (daemon operands) or configuration
+/// files. When composing diagnostics the value must not embed control
+/// characters, otherwise adversarial requests could smuggle terminal control
+/// sequences or split log lines. The helper replaces ASCII control characters
+/// with a visible `'?'` marker while borrowing clean identifiers to avoid
+/// unnecessary allocations.
+fn sanitize_module_identifier(input: &str) -> Cow<'_, str> {
+    if input.chars().all(|ch| !ch.is_control()) {
+        return Cow::Borrowed(input);
+    }
+
+    let mut sanitized = String::with_capacity(input.len());
+    for ch in input.chars() {
+        if ch.is_control() {
+            sanitized.push('?');
+        } else {
+            sanitized.push(ch);
+        }
+    }
+
+    Cow::Owned(sanitized)
+}
+
+fn format_bandwidth_rate(value: NonZeroU64) -> String {
+    const KIB: u64 = 1024;
+    const MIB: u64 = KIB * 1024;
+    const GIB: u64 = MIB * 1024;
+    const TIB: u64 = GIB * 1024;
+    const PIB: u64 = TIB * 1024;
+
+    let bytes = value.get();
+    if bytes.is_multiple_of(PIB) {
+        format!("{} PiB/s", bytes / PIB)
+    } else if bytes.is_multiple_of(TIB) {
+        format!("{} TiB/s", bytes / TIB)
+    } else if bytes.is_multiple_of(GIB) {
+        format!("{} GiB/s", bytes / GIB)
+    } else if bytes.is_multiple_of(MIB) {
+        format!("{} MiB/s", bytes / MIB)
+    } else if bytes.is_multiple_of(KIB) {
+        format!("{} KiB/s", bytes / KIB)
+    } else {
+        format!("{} bytes/s", bytes)
+    }
+}
+
