+++ crates/daemon/src/daemon/sections/module_definition.rs
@@ -1,0 +1,459 @@
+struct ModuleDefinitionBuilder {
+    name: String,
+    path: Option<PathBuf>,
+    comment: Option<String>,
+    hosts_allow: Option<Vec<HostPattern>>,
+    hosts_deny: Option<Vec<HostPattern>>,
+    auth_users: Option<Vec<String>>,
+    secrets_file: Option<PathBuf>,
+    declaration_line: usize,
+    bandwidth_limit: Option<NonZeroU64>,
+    bandwidth_limit_specified: bool,
+    bandwidth_burst: Option<NonZeroU64>,
+    bandwidth_burst_specified: bool,
+    bandwidth_limit_set: bool,
+    refuse_options: Option<Vec<String>>,
+    read_only: Option<bool>,
+    numeric_ids: Option<bool>,
+    uid: Option<u32>,
+    gid: Option<u32>,
+    timeout: Option<Option<NonZeroU64>>,
+    listable: Option<bool>,
+    use_chroot: Option<bool>,
+    max_connections: Option<Option<NonZeroU32>>,
+}
+
+impl ModuleDefinitionBuilder {
+    fn new(name: String, line: usize) -> Self {
+        Self {
+            name,
+            path: None,
+            comment: None,
+            hosts_allow: None,
+            hosts_deny: None,
+            auth_users: None,
+            secrets_file: None,
+            declaration_line: line,
+            bandwidth_limit: None,
+            bandwidth_limit_specified: false,
+            bandwidth_burst: None,
+            bandwidth_burst_specified: false,
+            bandwidth_limit_set: false,
+            refuse_options: None,
+            read_only: None,
+            numeric_ids: None,
+            uid: None,
+            gid: None,
+            timeout: None,
+            listable: None,
+            use_chroot: None,
+            max_connections: None,
+        }
+    }
+
+    fn set_path(
+        &mut self,
+        path: PathBuf,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.path.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!("duplicate 'path' directive in module '{}'", self.name),
+            ));
+        }
+
+        self.path = Some(path);
+        Ok(())
+    }
+
+    fn set_comment(
+        &mut self,
+        comment: Option<String>,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.comment.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!("duplicate 'comment' directive in module '{}'", self.name),
+            ));
+        }
+
+        self.comment = comment;
+        Ok(())
+    }
+
+    fn set_hosts_allow(
+        &mut self,
+        patterns: Vec<HostPattern>,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.hosts_allow.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!(
+                    "duplicate 'hosts allow' directive in module '{}'",
+                    self.name
+                ),
+            ));
+        }
+
+        self.hosts_allow = Some(patterns);
+        Ok(())
+    }
+
+    fn set_hosts_deny(
+        &mut self,
+        patterns: Vec<HostPattern>,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.hosts_deny.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!("duplicate 'hosts deny' directive in module '{}'", self.name),
+            ));
+        }
+
+        self.hosts_deny = Some(patterns);
+        Ok(())
+    }
+
+    fn set_auth_users(
+        &mut self,
+        users: Vec<String>,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.auth_users.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!("duplicate 'auth users' directive in module '{}'", self.name),
+            ));
+        }
+
+        if users.is_empty() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!(
+                    "'auth users' directive in module '{}' must list at least one user",
+                    self.name
+                ),
+            ));
+        }
+
+        self.auth_users = Some(users);
+        Ok(())
+    }
+
+    fn set_secrets_file(
+        &mut self,
+        path: PathBuf,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.secrets_file.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!(
+                    "duplicate 'secrets file' directive in module '{}'",
+                    self.name
+                ),
+            ));
+        }
+
+        let validated = validate_secrets_file(&path, config_path, line)?;
+        self.secrets_file = Some(validated);
+        Ok(())
+    }
+
+    fn set_bandwidth_limit(
+        &mut self,
+        limit: Option<NonZeroU64>,
+        burst: Option<NonZeroU64>,
+        burst_specified: bool,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.bandwidth_limit_set {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!("duplicate 'bwlimit' directive in module '{}'", self.name),
+            ));
+        }
+
+        self.bandwidth_limit = limit;
+        self.bandwidth_burst = burst;
+        self.bandwidth_burst_specified = burst_specified;
+        self.bandwidth_limit_specified = true;
+        self.bandwidth_limit_set = true;
+        Ok(())
+    }
+
+    fn set_refuse_options(
+        &mut self,
+        options: Vec<String>,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.refuse_options.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!(
+                    "duplicate 'refuse options' directive in module '{}'",
+                    self.name
+                ),
+            ));
+        }
+
+        if options.is_empty() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!(
+                    "'refuse options' directive in module '{}' must list at least one option",
+                    self.name
+                ),
+            ));
+        }
+
+        self.refuse_options = Some(options);
+        Ok(())
+    }
+
+    fn set_read_only(
+        &mut self,
+        read_only: bool,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.read_only.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!("duplicate 'read only' directive in module '{}'", self.name),
+            ));
+        }
+
+        self.read_only = Some(read_only);
+        Ok(())
+    }
+
+    fn set_numeric_ids(
+        &mut self,
+        numeric_ids: bool,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.numeric_ids.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!(
+                    "duplicate 'numeric ids' directive in module '{}'",
+                    self.name
+                ),
+            ));
+        }
+
+        self.numeric_ids = Some(numeric_ids);
+        Ok(())
+    }
+
+    fn set_listable(
+        &mut self,
+        listable: bool,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.listable.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!("duplicate 'list' directive in module '{}'", self.name),
+            ));
+        }
+
+        self.listable = Some(listable);
+        Ok(())
+    }
+
+    fn set_use_chroot(
+        &mut self,
+        use_chroot: bool,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.use_chroot.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!("duplicate 'use chroot' directive in module '{}'", self.name),
+            ));
+        }
+
+        self.use_chroot = Some(use_chroot);
+        Ok(())
+    }
+
+    fn set_uid(&mut self, uid: u32, config_path: &Path, line: usize) -> Result<(), DaemonError> {
+        if self.uid.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!("duplicate 'uid' directive in module '{}'", self.name),
+            ));
+        }
+
+        self.uid = Some(uid);
+        Ok(())
+    }
+
+    fn set_gid(&mut self, gid: u32, config_path: &Path, line: usize) -> Result<(), DaemonError> {
+        if self.gid.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!("duplicate 'gid' directive in module '{}'", self.name),
+            ));
+        }
+
+        self.gid = Some(gid);
+        Ok(())
+    }
+
+    fn set_timeout(
+        &mut self,
+        timeout: Option<NonZeroU64>,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.timeout.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!("duplicate 'timeout' directive in module '{}'", self.name),
+            ));
+        }
+
+        self.timeout = Some(timeout);
+        Ok(())
+    }
+
+    fn set_max_connections(
+        &mut self,
+        max: Option<NonZeroU32>,
+        config_path: &Path,
+        line: usize,
+    ) -> Result<(), DaemonError> {
+        if self.max_connections.is_some() {
+            return Err(config_parse_error(
+                config_path,
+                line,
+                format!(
+                    "duplicate 'max connections' directive in module '{}'",
+                    self.name
+                ),
+            ));
+        }
+
+        self.max_connections = Some(max);
+        Ok(())
+    }
+
+    fn finish(
+        self,
+        config_path: &Path,
+        default_secrets: Option<&Path>,
+    ) -> Result<ModuleDefinition, DaemonError> {
+        let path = self.path.ok_or_else(|| {
+            config_parse_error(
+                config_path,
+                self.declaration_line,
+                format!(
+                    "module '{}' is missing required 'path' directive",
+                    self.name
+                ),
+            )
+        })?;
+
+        let use_chroot = self.use_chroot.unwrap_or(true);
+
+        if use_chroot && !path.is_absolute() {
+            return Err(config_parse_error(
+                config_path,
+                self.declaration_line,
+                format!(
+                    "module '{}' requires an absolute path when 'use chroot' is enabled",
+                    self.name
+                ),
+            ));
+        }
+
+        if self.auth_users.as_ref().is_some_and(Vec::is_empty) {
+            return Err(config_parse_error(
+                config_path,
+                self.declaration_line,
+                format!(
+                    "'auth users' directive in module '{}' must list at least one user",
+                    self.name
+                ),
+            ));
+        }
+
+        let auth_users = self.auth_users.unwrap_or_default();
+        let secrets_file = if auth_users.is_empty() {
+            self.secrets_file
+        } else if let Some(path) = self.secrets_file {
+            Some(path)
+        } else if let Some(default) = default_secrets {
+            Some(default.to_path_buf())
+        } else {
+            return Err(config_parse_error(
+                config_path,
+                self.declaration_line,
+                format!(
+                    "module '{}' specifies 'auth users' but is missing the required 'secrets file' directive",
+                    self.name
+                ),
+            ));
+        };
+
+        Ok(ModuleDefinition {
+            name: self.name,
+            path,
+            comment: self.comment,
+            hosts_allow: self.hosts_allow.unwrap_or_default(),
+            hosts_deny: self.hosts_deny.unwrap_or_default(),
+            auth_users,
+            secrets_file,
+            bandwidth_limit: self.bandwidth_limit,
+            bandwidth_limit_specified: self.bandwidth_limit_specified,
+            bandwidth_burst: self.bandwidth_burst,
+            bandwidth_burst_specified: self.bandwidth_burst_specified,
+            bandwidth_limit_configured: self.bandwidth_limit_set,
+            refuse_options: self.refuse_options.unwrap_or_default(),
+            read_only: self.read_only.unwrap_or(true),
+            numeric_ids: self.numeric_ids.unwrap_or(false),
+            uid: self.uid,
+            gid: self.gid,
+            timeout: self.timeout.unwrap_or(None),
+            listable: self.listable.unwrap_or(true),
+            use_chroot,
+            max_connections: self.max_connections.unwrap_or(None),
+        })
+    }
+}
+
